[FIX] Folder의 id의 타입을 Long으로 변경

+++ src/main/java/com/mashup/molinkbackend/folder/Folder.java
@@ -15,7 +15,7 @@ public class Folder extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 40)
     private String name;
@@ -23,10 +23,10 @@ public class Folder extends BaseTimeEntity {
     @Column(length = 20)
     private String color;
 
-    private Integer parentId;
+    private Long parentId;
 
     @Builder
-    public Folder(String name, String color, Integer parentId) {
+    public Folder(String name, String color, Long parentId) {
         this.name = name;
         this.color = color;
         this.parentId = parentId;
